Delete LICENSE
Due to limitations of VS Code's handling of multi-root workspaces, it's not possible to hide this license file in the explorer view without hiding the license files for the code repos. This one just clutters up the explorer view, and since the files herein aren't part of the application itself, there's really no point in including this one.